:recycle::sparkles: turn Selection into component
